+++ tencentcloud/data_source_tc_cam_group_policy_attachments.go
@@ -109,7 +109,11 @@ func dataSourceTencentCloudCamGroupPolicyAttachmentsRead(d *schema.ResourceData,
 	groupId := d.Get("group_id").(string)
 	params["group_id"] = groupId
 	if v, ok := d.GetOk("policy_id"); ok {
-		params["policy_id"] = uint64(v.(int))
+		policyId, err := strconv.Atoi(v.(string))
+		if err != nil {
+			return err
+		}
+		params["policy_id"] = uint64(policyId)
 	}
 	if v, ok := d.GetOk("policy_type"); ok {
 		params["policy_type"] = v.(string)