+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ENHANCEMENTS:
 
 * Resource `tencentcloud_instance` support `SPOTPAID` instance. Thanks to @LipingMao ([#209](https://github.com/terraform-providers/terraform-provider-tencentcloud/issues/209)).
+* Resource: `tencentcloud_vpn_gateway` add argument `prepaid_renew_flag` and `prepaid_period` to support prepaid VPN gateway instance creation.
 
 BUG FIXES:
 * Fixed bugs that update operations on CAM POLICY do not work.
@@ -13,7 +14,6 @@ ENHANCEMENTS:
 
 * Resource: `tencentcloud_kubernetes_cluster` add support for `PREPAID` instance type. Thanks to @woodylic ([#204](https://github.com/terraform-providers/terraform-provider-tencentcloud/issues/204)).
 * Resource: `tencentcloud_cos_bucket` add optional argument tags
-* Resource: `tencentcloud_vpn_gateway` add argument `prepaid_renew_flag` and `prepaid_period` to support prepaid VPN gateway instance creation.
 * Data Source: `tencentcloud_cos_buckets` add optional argument tags
 
 BUG FIXES: